WW-5352 Move ParameterNameAware and ParameterValueAware
